# backend/docs/swagger.yaml
definitions:
...
  models.DocumentItem:
    properties:
      ballotId:
        type: integer
      ballotName:
        type: string
      committeeId:
        type: integer
      createdAt:
        type: string
      createdBy:
        type: integer
      createdByName:
        type: string
      description:
        type: string
      expectedAction:
        type: string
      expectedDate:
        description: keep as string if frontend sends YYYY-MM-DD
        type: string
      filePath:
        type: string
      id:
        type: integer
      meetingId:
        type: integer
      meetingName:
        type: string
      mimeType:
        type: string
      modifiedAt:
        type: string
      nNumber:
        description: N Number index
        type: integer
      projectId:
        type: integer
      projectName:
        type: string
      replaces:
        type: string
      status:
        description: To be notified / Notified / Draft / Deleted
        type: string
      subTypeCode:
        type: string
      subTypeName:
        type: string
      title:
        type: string
      typeCode:
        description: GENERAL / MEETING / PROJECT / BALLOT / RESOLUTION / FILE / FOLDER
        type: string
      typeName:
        type: string
      updatedByName:
        type: string
      version:
        type: number
    type: object
  models.DocumentListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.DocumentItem'
        type: array
    type: object
  models.DocumentLog:
    properties:
      action:
        type: string
      actionAt:
        type: string
      actionBy:
        type: integer
      actionByName:
        type: string
      actionByRoleID:
        type: integer
      actionByRoleName:
        type: string
      actionDetail:
        type: string
      documentId:
        type: integer
      id:
        type: integer
    type: object
  models.DocumentLogListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.DocumentLog'
        type: array
    type: object
  models.DocumentPathGroupsResponse:
    properties:
      data:
        items:
          type: string
        type: array
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.MaxNNumberResponse:
    properties:
      maxNNumber:
        type: integer
    type: object
...
  models.MessageResponse:
    properties:
      message:
        type: string
    type: object
  models.NotifyDocumentRequest:
    properties:
      body:
        description: 'Optional: if provided, send email'
        type: string
      subject:
        description: 'Optional: if provided, send email'
        type: string
    type: object
  models.NotifyDocumentResponse:
    properties:
      message:
        type: string
      sentTo:
        description: Number of email recipients (if email was sent)
        type: integer
      success:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/estandard/documents:
    get:
      consumes:
      - application/json
      description: Get list of documents with optional filters
      parameters:
      - description: Document title (partial match)
        in: query
        name: title
        type: string
      - description: Document type code
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DocumentListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Documents
      tags:
      - Documents
    post:
      consumes:
      - application/json
      description: Create or update a document and log action (CREATE / EDIT)
      parameters:
      - description: Document data (include id for update)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DocumentItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DocumentItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upsert Document
      tags:
      - Documents
  /api/v1/estandard/documents/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete document (set status = 'Deleted') and log action
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Document
      tags:
      - Documents
  /api/v1/estandard/documents/{id}/logs:
    get:
      consumes:
      - application/json
      description: Get all logs for a specific document
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DocumentLogListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Document Logs
      tags:
      - Documents
  /api/v1/estandard/documents/{id}/notify:
    post:
      consumes:
      - application/json
      description: Set document status to 'Notified' and log action. Optionally send
        email to committee members if subject and body are provided.
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Optional: Email subject and body to send notification'
        in: body
        name: request
        schema:
          $ref: '#/definitions/models.NotifyDocumentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NotifyDocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Notify Document
      tags:
      - Documents
  /api/v1/estandard/documents/max-n-number:
    get:
      consumes:
      - application/json
      description: Get the maximum n_number from documents with optional filters
      parameters:
      - description: Filter by committee ID
        in: query
        name: committeeId
        type: integer
      - description: Filter by meeting ID
        in: query
        name: meetingId
        type: integer
      - description: Filter by document type code
        in: query
        name: typeCode
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MaxNNumberResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Max N Number
      tags:
      - Documents
  /api/v1/estandard/documents/path-groups:
    get:
      description: Get distinct path prefixes (group by folder under documents/),
        e.g. documents/meeting, documents/resolution
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DocumentPathGroupsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get document path groups
      tags:
      - Documents
...
swagger: "2.0"